requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
baidusearch==1.0.3
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastmcp==2.2.10
Flask==3.1.0
frozenlist==1.6.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
html2text==2025.4.15
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.7.1
mdurl==0.1.2
multidict==6.4.3
openapi-pydantic==0.5.1
priority==2.0.0
propcache==0.3.1
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
python-multipart==0.0.20
Quart==0.20.0
requests==2.32.3
rich==14.0.0
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.7
sse-starlette==2.3.4
starlette==0.46.2
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uv==0.7.3
uvicorn==0.34.2
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.0
